MB-34753 Properly handle selectivity for LIKE predicates
There are two issues with this bug.

First, when a VALUED_SPAN is generated, currently we always produce a selectivity
for such spans. This is inconsistent with other spans where a selectivity is only
produced when valid statistics exists for the predicate. This is now fixed.

Second, for a LIKE predicate, the optimizer cannot generate an accurate selectivity
if there is a leading wildcard. Currently we always produce a NOT_AVAIL selectivity.
This causes problems since no cost/cardinality info is generated for such index scan.
The fix is to check whether valid statistics exists for the predicate, and if so,
produce a default selectivity instead of NOT_AVAIL such that cost/cardinality info
can be generated.

Change-Id: Id6178e3f238cb46c175ad98f697bc34b26e59224
Reviewed-on: http://review.couchbase.org/111830
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>