MB-34754 avoid inconsistent plans with CBO
Golang ranges through a map in undeterministic manner. In case
there are multiple index plans being compared and some of them
do not have valid cost associated with them, make sure we
compare cost for two plans that both have cost, and fall back
to legacy code if one or more of the plan is missing cost.

Change-Id: I2775e44ac87fb5cced645e552bd04a60f0ce8c16
Reviewed-on: http://review.couchbase.org/111879
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>