MB-35224 use n1ql ports for eventing functions UI MB-35102 add eventing err details
This change adds the ability to add ad hoc REST endpoints for specific purposes.
One such purpose is the upcoming javascript functions UI.

Also added infrastructure to report error details for javascript functions.
The details text is a bit muggy currently, so will enable the changes when it is cleaned up.

Change-Id: If62d1fcd0df805e5406251b0d5eb88bb8d38c110
Reviewed-on: http://review.couchbase.org/112464
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>