MB-34822 empty result set on reusing closed lcb handle
The underlying issue is that when an operator completes, we are sure that the indexing client will not be using 
the index connection. When it is stopped, the operator sends a stop to the indexing client and moves on,
but we have no guarantee that the indexing client has gone in a timely manner.

We could make the connection stop synchronous, but then if the IC hangs (and we have see those), then the whole request hangs.
I'm sure I could do a more thorough job of decoupling connection and operator, but however many defensive fixes I put,
it still gets me.
For now, we're going to do the same thing as we do for KILLED operators - send the operator to the GC.
Introduced a new _STOPPING status - if an operator does not clean up in time after a stop and doesn't transition to _COMPLETED,
we just ditch it. 

Change-Id: I7bc0e06b9a5629df3d4372b63ef4026037e4cd4b
Reviewed-on: http://review.couchbase.org/112603
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>