MB-34769 panic in parser on non existing UDF
The parser had encountered an error on a missing udf and should have stopped on the call to Error(),
but goyacc (unlike bison) continues on errors.

Added a Stop() lexer method to be called when we have errors that may later lead to undesired behaviour.
I suspect this is not the last time we'll call it.

Change-Id: Ib3fe726e3dc2d0a7a774c7e70c7b8a08aabce2de
Reviewed-on: http://review.couchbase.org/113094
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>