MB-35219 Fix cost/cardinality for LIKE predicate
LIKE predicate with trailing wild card such as name LIKE "Texas Wings%"
is transformed into a pair of range predicates: name >= "Texas Wings"
and name < "Texas Wingt". With distribution statistics available, it's
likely that both of these constants ("Texas Wings" and "Texas Wingt")
will reside in the same distribution bin, and since string type is
not numeric (for numeric we can do extrapolation based on boundary
values), the resulting selectivie for the pair of predicate ends up
equal to 0, which is misleading.

The fix is to make this a special case, where we remember that the
pair of predicates are originally derived from LIKE predicate, and
if the calculated selectivity is 0, re-calculate using name = <low>
to get a more realistic selectivity estimate.

Change-Id: I8708cb42c34db50b453ea5f09c869e3151137724
Reviewed-on: http://review.couchbase.org/113202
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>