MB-35555 fix n1ql test failures
Not clear why the error_cases test isn't failing as a matter of norm, because two errors expected have always been wrong.
Have changed the parser Stop() method not to return a syntax error in error, and have asked Isha to check the other two errors.

The hang and / or panic were coming from operators again being marked as completed in a race condition. 

Change-Id: I061049c1f239d4b434fc82f9f9c918248f07bacf
Reviewed-on: http://review.couchbase.org/113287
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>