MB-35811 connection pool leak querying system:indexes
The underlying issue is that with the fix to MB-33185 we have relied on bucket
finalizers to close connection pools, but somehow there's now a reference to the
bucket being closed, which means that the finalizer never gets used.

I have taken a different approach in the datastore/couchbase package: rather than
having one refresh function that does both a refresh and reload and always creates
a new pool, I have split it in two, one for refresh (used when we are looping through
keyspaces) and one for reload (when things have gone wrong and we reload from scratch).

Reload is mostly unchanged (except for manually closing each known bucket), refresh uses the current pool and just checks if buckets exist.
In so doing, it does not create new connections pools, and there's no closing needed.
This makes the refresh faster.
If things have changed, refresh just gives up and does a full reload.

Change-Id: I51ae48f41d8ecd94c29fcfc2a0245951c65859f2
Reviewed-on: http://review.couchbase.org/114748
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-on: http://review.couchbase.org/114842
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>