MB-32994 improve projection throughput MB-36401 improve filter instantiation MB-36655 lower cost of kept-alive operators MB-36829 use simpler monotonic clock for profiling and maintenance
There's a lot to chew here, but the changes are all related to speeding up execution.

Most complicated is the first.
Essentially, currently projection is handled by two different operators, which requires to instantiations,
two destructions, sending values across an extra hop, and the creation of 2 annotated values and a scope value.
We eliminate the final projection altogether (avoid one instantiation, one destruction, and one hop),
and instead we pass along the projection, but with the original value tagged along, for the benefit of the
order operator (which in turn saves an annotated value most times).
Final Projection is accepted (and produced) in plans for backwards compatibility (think mixed version clusters),
but ignored.

Next related step, is to alter the sequences in the planner to skip creating sequences if the first child is
a sequence itself, in which case all the other children are appended to the first child.

This saves a sequence.
Next related step is to hack the planner to attach the stream first, and then authorize.
This saves another sequence.

The plan has now switched from Sequence((Authorize, Sequence(Scan, Fetch, Sequence(Initial, Final)), Stream) to
Authorize Sequence(Scan, Fetch, Initial, Stream)

The next related step is not to fork setup operators like Authorize, Sequence, Parallel, but to
execute inline in the parent.
This saves another two goroutine forks, which in turn saves allocating and garbage collecting two
stacks.
This adds two more profilings because we have not to count inlined time in the parent execution.

Each of these changes improves Q1 by about 3K/sec, so cumulatively Q1 goes from 127k, 4ms to 140k 2ms.

Next related step is to lower profiling time by moving away from golang time.Now() (which does two syscalls,
one for wall time, which we don't need, and one for monotonic time).
We create a simpler util.Time type which carries only the monotonic time, and use it wherever we just need to
measure durations.

This lowers latency from 2ms to 1.75ms.

I wish Q2 was better, but it only goes up from 67k/sec to 70.


Change-Id: I32c20c78eaff6971c566f08c7c88f8167e6be74a
Reviewed-on: http://review.couchbase.org/116356
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>