MB-36653 CBO: drop index statistics when an index is dropped
When indexer information is being loaded (i.e. a query is being run), check
and potentially remove indexes that has been dropped from the dictionary.
This check is being done every second or so (checks which are within a second
of previous checks are ignored). When an index is dropped from dictionary
cache it is also removed from metakv.

Change-Id: Ic0995b3958e693524c260ffef3d51a13755a163f
Reviewed-on: http://review.couchbase.org/117156
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>