MB-37038 Combine Sequence operators at run time instead of plan generation time
A previous performance optimization (MB-36655) combines Sequence operators
if a new Sequence operator has another Sequence operator as its first child.
In such cases we don't generate a new Sequence operator, instead, we tag
the children of the new Sequence to existing children of the existing
Sequence operator. This saves a Sequence operator.

The problem with this approach is that, in the planner sometimes we need to
generate temporary plans for consideration only. For example, with CBO we
generate both a hash join plan and a nested-loop join plan, and then pick one
based on cost. However, if as part of this temporary plan generation we
already modified the children array of an existing Sequence operator, then
this change is no longer "temporary" and may affect subsequent plan execution.
In case of the Merge test, we generate a hash join plan, and as part of that
we extend an existing Sequence operator's children array with an Alias
operator. Subsequently we decided to throw away this plan and generate a
nested-loop join plan instead. Later on we end up tagging along the same
Alias operator again, resulting in two Alias operators on top of each other.
This creates problems at runtime.

Moving forward temporary plan generation will be more prevalent especially
when we start to rearrange join ordering. To avoid issues like this, we
"move" the optimization of combining Sequence operators from plan generation
time to execution time. During plan generation we still generate potentially
redundant Sequence operators (and thus avoid modifying existing Sequence
operators). Then once a final plan is chosen, and when we create an
execution tree from a plan tree, as a Sequence operator is being generated,
we check whether the first child of it is also a Sequence operator, and
if so, we tag along the new children of the new Sequence onto the children
array of the existing Sequence operator, this way the same optimization
as previou is still done, except it's done at execution time instead of
plan generation time.

Change-Id: Ice6655de42fab859b306b62979bf53372c3915e2
Reviewed-on: http://review.couchbase.org/118400
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>