MB-37082, MB-27595: Updated HTTP2 setup to error on invalid configurations.
The current implementation of HTTP2 setup can cause invalid configurations
to be created, the behavior exhibited from this misconfiguration is that
conformant HTTP2 clients will trigger a connection-fatal error whenever
they try to connect.  This change uses Go's built-in validation logic for
http2 to fail startup if an invalid TLS configuration is provided.

Change-Id: I0d71e8b77401d4eb2b5d9a5b4d50b597934af6b7
Reviewed-on: http://review.couchbase.org/c/query/+/118613
Tested-by: Isha Kandaswamy <isha@couchbase.com>
Reviewed-by: Isha Kandaswamy <isha@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>