MB-37493 use single deadline for http request
We have used two http request deadlines (header + read) to be handle DOS attacks
but be more lenient when reading request bodies.

It turns out that the net package is pretty poor at handling deadlines and this can be quite a substantial bottleneck.
So we are switching to a more stringent deadline set up (5 seconds for header and body, rather than 5 seconds for
header and another 30 seconds for body), because that gives us an extra 4k throughput in Q1 and cuts 99% latency from
1.6ms to 1.2ms  

Change-Id: I5f5385acf4314bd12087cdab018a499c6763058b
Reviewed-on: http://review.couchbase.org/120404
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>