MB-37602 Incorrect span generated if index key expression has AND condition
Predicates in the WHERE clause is broken up on AND bounderies into
individual filters. This can be problematic when sarging index keys
if an index key expression has AND condition in it, since we use
each individual filter for sarg purpose.

The fix is to detect this unusual situation, and use the entire expression
instead of individual filters for sarging index keys.

Change-Id: I6fcaa32a5f53e6c3a5b16644d6477b2743adec7a
Reviewed-on: http://review.couchbase.org/122140
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>