MB-37602 Incorrect span generated if index key expression has AND condition
Predicates in the WHERE clause is broken up on AND bounderies into
individual filters. This can be problematic when sarging index keys
if an index key expression has AND condition in it, since we use
each individual filter for sarg purpose.

The fix is to detect this unusual situation, and use the entire expression
instead of individual filters for sarging index keys.

Change-Id: Ia6fa3e9a6e4cce387b7cfd9f8aac470db7bfdd92
Reviewed-on: http://review.couchbase.org/122142
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Wayne Siu <wayne@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>