MB-38105 Index scan may generate incorrect span when UNNEST is involved and unnest scan is not feasible
When a query involves UNNEST the planner considers UNNEST SCAN as part of
its consideration, and for that to work properly, the planner includes
predicates on the unnest alias to the predicates array of the primary
keyspace. In case an unnest scan is not feasible (lack of appropriate
arry index), the planner may choose an index on the primary keyspace, and
if a predicate involves both the primary keyspace and an unnest alias,
it may be used incorrectly as part of index span and causes issues
at runtime since the unnest alias is not yet available.

There is already logic (as part of ANSI JOIN implementation) to detect
that a span should not contain any references to the keyspace being
considered. This logic is enhanced to also detect any references to
unnest aliases as well. For this to work a new map of all unnest aliases
is now kept on the primary keyspace.

Change-Id: I3a0afe40bbfaf13139e0fac8808cf19e1ba80691
Reviewed-on: http://review.couchbase.org/123040
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>