temporary fix for GetRandom() until get random entry for collections is implemented
GetRandom() has not been maintained in a while by the KV team, since they weren't expecting anyone to use it.
It will be revamped for collections, but until we know what the actual implementation is, this is the
simplest fix for the default collection which doesn't affect any other team.

We will revisit this once we know what GetRandom() looks like in the KV.

Change-Id: Ifafd4126701f615a6a25d4ddcc1116905ee4458e
Reviewed-on: http://review.couchbase.org/123754
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Chang Liu <chang.liu@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>