Collect cluster settings for IA
For ease of management now the clustering configstore does not only report the local node name and state, but also a pointer to the local node cluster info.
The cluster in turn now has a new Settings() method reporting a system specific map[string]interface{} of settings.

For couchbase we currently only report memory quotas - but this is the vehicle to report other interesting stuff, once we need it.

To use
1st in cbq_engine/main.go store the configstore pointer in the package where you need it, much  like we do for monitoringInit() for the distributed
package.
Unluckily I cannot expose this through distributed because of circular references
2nd get the local cluster from the configstore
3rd if there is a cluster defined, get the settings

from there drill into Settings["memory"["index"], eg:

        var settings map[string]interface{}
        var memory map[string]interface{}

        cluster, _ := configstore.Cluster()
        if cluster != nil {
                settings, _ = cluster.Settings()
        }
        if settings != nil {
              memory = settings["memory"]
        }
        if memory != nil {
                 indexQuota  =memory["index"]
        }

Change-Id: Ifd1e2507ae49008edc27ed32b35ff63477ec21f7
Reviewed-on: http://review.couchbase.org/124067
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>