MB-38484 metrics, pools and caches a bottleneck again.
Sharding caches and pools over 16 buckets no longer sufficient with the latest
performance improvements - need to shard over the # of CPUs.
Also the sample metric is a bit of a bottleneck again - streamlined the code
to hold the lock as little as possible, and rather than having a sample pop and then a push
(we shrink and then append to the sample slice), created a more streamlined method
that just rotates the old value out and populates the new value on the now empty top.

This change adds another 8k to Q1.
Also, with a streamlined version of Q1 of just 100 clients, I now hit 100k/sec on 0.3ms latency!

Change-Id: Ic23cfcd745ac59dfdc4771b107de19705d8b8bf5
Reviewed-on: http://review.couchbase.org/c/query/+/124513
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>