MB-37115 Make sure lastOp is always properly updated
CBO sometimes needs to know what is the last operator generated in order
to properly cost the current operator. The lastOp field in builder struct
should always be updated when a new operator is generated for the query
plan. We now make sure that whenever a new operator is appended to either
subChildren or children slice of builder structure, the lastOp field is
properly updated. Also added a new function to handle adding a parallel
operator.

Change-Id: I537926396d1f88b368ca42f7814c0c26a2282688
Reviewed-on: http://review.couchbase.org/c/query/+/124542
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>