MB-38564 wrong result possible with left join and IS MISSING predicate is present on inner table
For ANSI OUTER JOIN the filters in the WHERE clause should be evaluated
after the join is done (post join) and thus should not be pushed to be
evaluated at the inner scan. Certain filters are safe to push down, e.g.,
if a filter can eliminate NULL-extended rows from the result of an outer
join (null-rejecting filters). In this defect, an IS MISSING filter
(which is NOT null-rejecting) is being incorrectly pushed down to inner
scan, and an incorrect MISSING span is generated for the inner index
scan, leading to wrong result.

The fix is multi-fold:

  1. When WHERE clause expression is being classified, individual filters
     are still put in the filters array of BaseKeyspace. This behavior
     remains the same.
  2. When a join is considered, if it is an outer join, then only
     consider filters marked as ON-clause filter for index selection
     purpose for the inner scan.

Change-Id: Ia48a3f55214083b6d07120225786d932156f4c20
Reviewed-on: http://review.couchbase.org/c/query/+/124952
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>