MB-38968 Incorrect ANSI OUTER JOIN to ANSI INNER JOIN transformation
The problematic query has an ANSI OUTER JOIN in the form of l LEFT JOIN r
and the WHERE clause has an IN-clause in the form of:
  l.field IN (CASE WHEN r.filed IS VALUED THEN ... ELSE ... END)
Current code incorrectly treat the IN-clause filter as a null-rejecting
filter and transformed the ANSI OUTER JOIN to ANSI INNER JOIN.
For an IN clause to be null-rejecting with regard to keyspace r, the
left-hand side of the IN-clause must reference keyspace r.

Fixed the logic for determination of null-rejecting criteria such that
a subexpression must reference the relevant keyspace as a field or an
identifier. Then the null-rejecting property of certain expressions
such as EQ, IN, etc can be determined.

Change-Id: I9b7a1f1b43a57d102198db260b54233489b17c1a
Reviewed-on: http://review.couchbase.org/c/query/+/126638
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>