MB-38801 concat() with 3 arguments is not covering
When string concatenation is done using '||', the parser
generates a Concat expression for each '||' encountered,
and thus something like "a" || "b" || "c" will have nested
Concat node where the top Concat has another Concat as its
first child. In contrast, when the concatenation is done
using concat("a","b","c"), we have a single Concat
expression with 3 children. This looks different than the
nested Concat expression and thus in covering determination
we treat them differently.

The fix is to be a bit smarter when a Concat node is being
created, and check whether it's first child is already
a Concat expression, and if yes, simply extend the
operand list of the Concat expression instead of creating
a nested Concat expression.

The same is also done for Concat2, although it's a much
rarer occurance for Concat2, you need to explicitly
specify something like Concat2("-", Concat2("-", "a", "b"), "c")
for this new optimization to kick in.

Change-Id: Id3cee3c9a62d9ac1ebe63259819e01dbba1e3663
Reviewed-on: http://review.couchbase.org/c/query/+/126734
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>