MB-38996 Queries with correlated subquery in WHERE clause may return incorrect result with CBO turned on
When CBO is turned on, selectivity is calculated for each individual
filter. As part of this process, we'll try to look for a histogram
for an expression. Part of this process involves stripping keyspace
reference from an expression in order to match a histogram. A copy
of the expression is made before this stripping process. However,
if the expression contains a correlated subquery, when the expression
is copied, the subquery itself is NOT copied. As a result, we end
up stripping keyspace reference of the original subquery expression,
resulting in error at runtime.

Since there is no way a subquery expression can match a histogram,
we skip subquery expression when stripping keyspace references.

Change-Id: I6c38cdbe8e9d4b5c1fe63efba0afcdc8b4d233b7
Reviewed-on: http://review.couchbase.org/c/query/+/126758
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>