MB-38909 request REST parsing bottlenecks
Although this falls under TCO, the aim here is saving memory, so this also falls under memory quota.
On Q1 this code results in overall memory usage going from 3.5GB to 2.75GB.

There are a few things at play here.
1- compiler escape analysis for memory allocation has changed, and putting a structure on a frame no longer works.
2- we are building and populating maps and ditching them within the REST API request parsing code, which needlessly puts the GC under pressure and increases average memory consumption.
3- we are parsing url encoded forms even if the request is a json.

Re 1- this is an initial change. I don't know if I can improve on the httRequest{}, urlArgs{} and jsonArgs{} allocation, because they get pointed to by an interface variable, which means,
as far as I can find on the internet, that the escape analysis will always place them on the heap.
With this change though, I do save an allocation.
I will play with pools and returning values and not pointers as time allows.

Re 2- what we are doing is scanning the json document, rather than parsing it.
This means that we do not create a map and clean it straight away.
As an added saving, we use a static array for the initial parameters, normal parameters and even parameters that have to be accessed directly, meaning
we don't have to allocate slices either.

I'd like to do the same for urlArgs, but I will need to write code to scan url forms.

Re 3- we only scan url forms for urlArgs.

So for json args we only allocate a map for named parameters, and everything else is static.
That's a saving of 5 maps, which leads to less GC cleanup, which leads to less memory.

The side benefit of all of this is that scanning a json is faster than decoding it,
so Q1 throughput goes from 149K @ 1.2ms  to 153K @ 0.8ms.

Change-Id: I88d1250db428ca387923488dbfff8581f210c7d5
Reviewed-on: http://review.couchbase.org/c/query/+/126912
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>