MB-37115 Add infrastructure for a new optimizer package
To prepare for upcoming join enumeration and more cost-based optimizer
functionalities, we need to add a new optimizer package to be able to
import planner package for building individual plan operators. The
existing optimizer package is renamed to optutil since it currently
contains utility functions for calculating selectivities and costs.
A new Optimizer interface is defined in the planner package to represent
the new optimizer, which is now separate from the existing planner
such that we no longer rely on the planner to build the skeleton
of a query plan. A pointer for the new optimizer is also added to
planner context and execution context such that when subquery
is being optimized we'll have the proper handle for the new optimizer.

Change-Id: Ia25b3163311f0bd463a005d56025491991c60fc1
Reviewed-on: http://review.couchbase.org/c/query/+/126968
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>