MB-38469 initial operator stack too small
This is another memory usage issue: essentially, the initial stack for goroutines is too small,
so short running go routines spend an inordinate amount of time allocating new stacks from
the heap and copying it.
This leaves a lot of empty blocks that need to be cleaned up.
Our short running goroutines are all the operators that we fire, for Q1 at a rate of 500K per second.
From some memory profiling I have done, each operator reallocates stacks an average of 4 times, which leaves
2M blocks per second to be cleaned.

This is a well known go issue that has been going on since 2016.

I have tried two approaches described in the go issue: 1- sizing the stack immediately and 2- having a pool of goroutine workers.
1- seems to work better: we only resize the stack once and, since the goroutine has just started and there's very little on the frame,
there's very little to copy
2- avoids all the stack allocations, but at the cost of keeping a pool of stacks, which doesn't really save memory.
So the actual fix is in base.go - the new file is a low cost pool of goroutines, which might be handy later, if we want to implement some
workers based thing.

Change-Id: Id8cef1b9d600d042b05485616f051cb290236e14
Reviewed-on: http://review.couchbase.org/c/query/+/127148
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>