MB-39054 extra allocations from value and expression
This is an initial bit of work for memory quota.
The issue here is endemic across expression and value, the work here is the
bare minimum that's needed for Q1.

The first issue - the whole Applied interface for function base requires garbage collector work
because the variadic list of parameters is in fact a dynamic slice that gets silently allocated from the hep, and cleaned.
For cons_array I used the same code as in cons_object, ie ditch the Applied interface.
For other functions, we need to find a better way, like maybe have two methods, one variadic, one not.

Two - it's best two size maps if we know a likely size, because this leads to one allocation for both map structure and buffer,
rather than two (one for the map, one later for the buffer when we start assigning elements).

Lastly - byte slice constants get escaped to the heap because the compiler is not smart enough to figure that they won't change.
So we are better off use the StringWriter interface.
Issue is, this is defined in 1.12, and the indexer compiles with 1.11, so using it directly breaks builds.
The workaround is to cast the writer back to a buffer writer - since this is what we are using to start with.
We'll ask indexing to upgrade, and then cleanup better.

By making sure that things don't get escape, we lower overall memory allocation.
The side effect of this is that with lesser mallocgc and GC work, we get a little better
with throughput.
In this case, 2K req/secs.

I plan to do all of this cleanup little by little together with MB 39084, cleanup value and expression pools.


Change-Id: I0980a9c2633b5e0de9ce068544762ac51c7a0456
Reviewed-on: http://review.couchbase.org/c/query/+/127149
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>