MB-39236 In correct result from hash join if CBO switches build and probe side
When CBO is turned on, it considers both sides of a hash join to be
the build/probe side, and its possible that the build/probe sides
can be switched. If this happens, current code is not properly
adjusting all necessary fields for the switch, leading to runtime
problems.

The fix is to move the setup of build/probe side info after
the build side is determined.

Change-Id: Id8e75891d870fe895e4cefe25d509472dcdc2176
Reviewed-on: http://review.couchbase.org/c/query/+/127560
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>