MB-39275 Properly calculate join cost for outer join
For inner join, the ON-clause is preprocessed before join is
being considered. For outer join, the ON-clause is processed
at the point of the join. Part of this processing is to mark
the keyspaces that's already been planned as "done", and thus
not include these keyspaces in the keyspacemap of the filter,
such that these join filters can be used as index filters of
the inner scan. However when we calculate selectivity for
these join filters, we incorrectly treat them as table filters
due to lack of other keyspace references, and the selectivity
is not properly calculated.

The fix is to do selectivity calculation before we remove
references to keyspaces that are considered "done".

Change-Id: I561cee9316c3e32ff46c7eb29f057edaf545cd2f
Reviewed-on: http://review.couchbase.org/c/query/+/127681
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>