MB-39296 Unnest scan not working properly in presence of tokens index
This is a regression of 2nd part of fix for MB-35871.

When unnest scan is considered, it needs predicates on the unnest
alias. These unnest predicates are added by function addUnnestPreds().
Previously these unnest predicates are added right before we
consider unnest scan. As part of the fix for MB-35871 this is
moved up earlier. The issue with this defect is in presence of
tokens index, we add new predicates for the tokens index, and
as part of that process, we reclassify the predicates, and
"lost" the unnest predicates in this process.

The fix is to add a call to addUnnestPreds() after we add new
predicates and reclassify the predicates.

Change-Id: I5cd5230f12c89124e460298ebbceae841e1671b0
Reviewed-on: http://review.couchbase.org/c/query/+/128032
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>