MB-39443 Incorrect cardinality estimate for queries with join and local predicates
For an index scan the selectivies for index filters are
already applied for at the index scan level and thus
should NOT be reapplied at the Filter operator level.
This is achieved by marking filters with a flag when
an index scan is being created. This works previously
since for each keyspace an index scan is created once.
With CBO we now consider alternative join methods for
each join, as a result for the inner keyspace an index
scan may be considered multiple times, when different
join methods are being considered. As a result the
index flags set on the filters may be cleared when
new index scans are being considered. This results in
selectivies potentially being applied multiple times.

The fix is to move the logic for marking index flag
on the filters from when an index scan is being
created, to after a join method has been finalized
(therefore no new index scan will be considered for
this keyspace again).

Change-Id: I25bb119a5a8d7d0182e1c1500c3e4e15e8f94394
Reviewed-on: http://review.couchbase.org/c/query/+/128435
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>