MB-32869 RBAC for functions
so with this one, users get granted roles like
query_manage_functions or query_execute_functions.
We already add the grant / revoke of functions RBAC implemented, so this change
mostly deals with using the roles and checking the permissions.
There's two extra things - first functions RBAC requires roles at the scope level.
This wasn't implemented so I added it.
The second is that GRANT / REVOKE actually used keyspace_ref rules, which allowed for the use of options and aliases.
I removed those.
The third is that semantics should check that the roles and objects match,
like you should be able to grant admin on a collection or scope - valid options are nothing or bucket
for execute and manage functions the options are global, bucket, scope but not collection.
This change is not specific to functions, so I have logged a different MB.

Change-Id: Ibc2aa434b64ad5d59861da91dc33d9ee91fcd095
Reviewed-on: http://review.couchbase.org/c/query/+/129942
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>