MB-39785 Predicate pushdown - part1: push selection filter to Fetch, IndexScan3 and ExpressionScan operators
Currently the WHERE clause is evaluated at Filter operator which is added
after all scans and joins. In case we have a filter on non-indexed field
this filter will not be applied until all joins are done, which means
we could be doing a lot of useless join work.

Part 1 of the fix pushes the selection filters down to either the Fetch
operator (for non-covering index scans) or the IndexScan3 operator
(for covering index scans). Filters are also pushed down to ExpressionScan
operator. If a subquery term is used, an explicit Filter operator is added
after the plan operators for the subquery and the Alias operator.

This way the filters will be evaluated early on before joins. Once
a filter is pushed down, it will be removed from the filter to be
applied on the Filter operator. Note this removal only happens
when there is an exact match. For derived filters or transformed
filters, these will be pushed down, however the original filter is
still applied at the Filter operator.

Change-Id: Ib0432a697c88c0c3e82a725baa00341a574b4954
Reviewed-on: http://review.couchbase.org/c/query/+/129967
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>