MB-39785 Predicate pushdown - part2: push selection filter to join operators
Part 2 of the fix pushes the selection filters down to join operators
including NLJoin, HashJoin, NLNest, HashNest and Unnest. The pushed
down filters are evaluated after the join operations (postjoin)
which is equivalent to evaluating these filters with a Filter operator
after the join. The advantage of evaluating these filters at the join
operators is in case there are multiple joins, we could potentially
eliminate results in earlier joins before later joins are processed.

Change-Id: Ie71755a18e633583b8434841cfc0a9f22ee09ec2
Reviewed-on: http://review.couchbase.org/c/query/+/130064
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>