MB-39785 Predicate pushdown - part3: cost and selectivity adjustments
Part 3 of the fix adjusts selectivity and cost for pushing down
filters to various operators. Since we do need to properly
account for only those filters whose selectivity has not been
applied, we need to temporarily mark the filters used in
index scan and hash join such that we don't double-apply the
selectivities. For joins, we gather the filters before any
plan is considered, thus once we have a plan, we need to
adjust for selectivities first before applying them.

Change-Id: I67e08a10edd30a72771eb001fc078f38b47f9c77
Reviewed-on: http://review.couchbase.org/c/query/+/130170
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>