MB-37115 Better matching of index expressions when index scan has multiple spans
When an index scan is generated the optimizer uses the index spans
to try to match the original expressions used to generate the index
spans such that selectivities of these filters are not applied
multiple times. In case there are multiple spans, resulting either
from either IN-clause, or OR-clause, or combination of the two,
we can now properly generate the appropriate expression based on
the index spans, in order to properly match with the original
expressions used to generate the index spans.

This checkin only covers IN-clause and OR-clause. Other situations
where multiple spans can be generated (such as != clause) is not
handled.

Change-Id: I46b91b252b1eeec1fe49a0b0038d5d0d6ac0a020
Reviewed-on: http://review.couchbase.org/c/query/+/130360
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>